collecting and getting debuggers from new debugger module
